Refactored Vulkand Backend, small bugs fixes and minor improvements

+++ engine/src/platform/Windows/WinFile.cpp
@@ -10,7 +10,7 @@ b8 readFile(const std::string &path, hk::vector<u8>& out)
     // ALWAYS_ASSERT(file.is_open(), "Failed to open a file:", path);
     if (!file.is_open()) { return false; }
 
-    const u32 size = file.tellg();
+    const u32 size = static_cast<u32>(file.tellg());
     file.seekg(0, file.beg);
     out.resize(size);
     file.read((char*)(out.data()), size);
